Refactor launcher.py into modular app package
Break up the large monolithic Launcher.py into a modular package:

- add app.py, config.py, process_tracker.py, tab_manager.py, update_checker.py, a /tabs directory (base, command_line_tab, markdown_tab, perf_tab, script_tab).

- Move utilities (ordered_logger, parse_ansi, utils) into Launcher/LauncherScript/_lib.

- Launcher.py - now lightweight entry point
