+++ Launcher/LauncherScript/_lib/__init__.py
@@ -1,0 +1,13 @@
+# _lib/__init__.py - Internal launcher utilities
+
+from .parse_ansi import AnsiParser
+from .ordered_logger import OrderedLogger
+from .utils import RingMovingAverage, is_shift_held, ensure_dependencies
+
+__all__ = [
+    'AnsiParser',
+    'OrderedLogger',
+    'RingMovingAverage',
+    'is_shift_held',
+    'ensure_dependencies',
+]
