+++ Launcher/LauncherScript/_lib/parse_ansi.py
